requirements.txt
opencv-python==4.5.5.64
numpy==1.22.3
pyserial==3.5
PyQt5==5.15.4
tensorflow==2.8.0